Updated file structure, with CMake included, cmake needs refinement and has not been tested. Also made module structure for all files.
