+++ src/Satisfaction/SatisfactionState.h
@@ -1,7 +1,7 @@
 #ifndef SATISFACTIONSTATE_H
 #define SATISFACTIONSTATE_H
 
-#include "../City.h"
+#include "managers/City.h"
 
 #include <string>
