+++ src/command/AllocateResourcesCommand.cpp
@@ -1,5 +1,5 @@
 #include "AllocateResourcesCommand.h"
-#include "City.h"
+#include "managers/City.h"
 
 AllocateResourcesCommand::AllocateResourcesCommand(City* newCity){
     city = newCity;
