+++ src/command/SetPolicyCommand.cpp
@@ -1,5 +1,5 @@
 #include "SetPolicyCommand.h"
-#include "City.h"
+#include "managers/City.h"
 #include "Policy.h"
 
 SetPolicyCommand::SetPolicyCommand(City* newCity){
